Create README.md
Pac-Man is a popular arcade game of the times of slot machines. Its rules are very simple: you need to eat all the points on the game field and at the same time not to get caught by ghosts which can be neutralized for some time with the help of energizers.